Move to_camel_case helper to util module.
